pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "aiosqlite>=0.21.0",
    "asyncpg>=0.30.0",
    "discord-py[voice]>=2.4.0",
    "flask-login>=0.6.3",
    "flask-wtf>=1.2.2",
    "matplotlib>=3.10.0",
    "oauthlib>=3.2.2",
    "pynacl>=1.5.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.0.1",
    "slack-sdk>=3.34.0",
    "trafilatura>=2.0.0",
    "twilio>=9.4.5",
    "types-python-dateutil>=2.9.0.20241206",
]
